Move console log spec and test util back into libs/common
