+++ libs/logging/src/index.ts
@@ -1,4 +1,3 @@
 export { LogService } from "./log.service";
 export { LogLevel } from "./log-level";
-export { interceptConsole, restoreConsole } from "./intercept-console";
 export { ConsoleLogService } from "./console-log.service";
