--- libs/logging/src/intercept-console.ts
@@ -1,20 +1,0 @@
-const originalConsole = console;
-
-declare let console: any;
-
-export function interceptConsole(): {
-  log: jest.Mock<any, any>;
-  warn: jest.Mock<any, any>;
-  error: jest.Mock<any, any>;
-} {
-  console = {
-    log: jest.fn(),
-    warn: jest.fn(),
-    error: jest.fn(),
-  };
-  return console;
-}
-
-export function restoreConsole() {
-  console = originalConsole;
-}
